+++ tests/Unit/Mvc/Controller/Traits/Fixtures/Models/Error.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * This file is part of the Phalcon Kit.
+ *
+ * (c) Phalcon Kit Team
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace PhalconKit\Tests\Unit\Mvc\Controller\Traits\Fixtures\Models;
+
+use Phalcon\Mvc\Model;
+
+final class Error extends Model
+{
+}